# doc/swagger.yaml
openapi: 3.0.3
info:
  title: Sky's Auth System
  description: This is the authentication system for Sky's webservers and websites.<br>
    Some endpoints require "Super Admin" authentication. These will be marked as such, 
    and can only be used by users with the appropriate permissions. These generally
    include the project endpoints, and some administration endpoints.
  contact:
    email: auth@skystuff.cc
  version: 1.0.0
servers:
  - url: https://auth.skystuff.cc/api/
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /user/get/:
    get:
      tags:
        - Users
      summary: Get the currently signed in profile.
      description: Get details about a user.
      operationId: user.get_user
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing Authorization.
  /user/keys/:
    get:
      tags:
        - Users
      summary: Get all the API keys of a user.
      description: Get all the API keys of a user, returns as a JSON list.
      operationId: user.get_user_keys
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: The total number of users to list. Maximum of 50.
        - in: query
          name: offset
          schema:
            type: integer
          description: Offset for search results.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keys'
        '401':
          description: Missing Authorization.
  /user/projects/:
    get:
      tags:
        - Users
      summary: Get all the projects a user has access to.
      description: Returns all of the projects a user has access to as a JSON list.
      operationId: user.get_user_projects
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: The total number of users to list. Maximum of 50.
        - in: query
          name: offset
          schema:
            type: integer
          description: Offset for search results.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projects'
        '401':
          description: Missing Authorization.
  /user/projects/all/:
    get:
      tags:
        - Users
      summary: Get all the projects a user has access to.
      description: Returns all of the projects a user has access to, and all other public projects and the user's approval status.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: The total number of projects to list. Maximum of 50.
        - in: query
          name: offset
          schema:
            type: integer
          description: Offset for search results.
      operationId: user.get_user_projects_all
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatuses'
        '401':
          description: Missing Authorization.'
  /user/tokenrefresh/:
    post:
      tags:
        - Users
      summary: Reset normal and secure token for a user account.
      operationId: user.refresh_token
      description: This will refresh the tokens on a user account, in the event that one has been leaked. This requires any token, even the not priviledged one. The SecureAuthentication is the user's password, so to reset that the user must change their password.
      responses:
        '200':
          description: Successful operation.
        '401':
          description: Unauthorized.
  /user/login/:
    post:
      tags:
        - Users
      summary: Log in to an account.
      operationId: user.login
      description: This will also set 2 cookies, 'Authentication' which is the normal user.token, and 'SecureAuthentication', which is the user's hashed password. The 'SecureAuthentication' only lives for 15 minutes, whereas the 'Authentication' lives for 1 week.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.login'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Missing body element. Username already registered. Username too long/short.
  /user/create/:
    post:
      tags:
        - Users
      summary: Create a user.
      operationId: user.create_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.create'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Missing body element. Username already registered. Username too long/short.
  /user/edit/:
    patch:
      tags:
        - Users
      summary: Edit a user.
      description: If editing the username or password, both must be passed due to how hashing passwords work.
      operationId: user.edit_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.create'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Password provided but not username, or vice versa. Username already registered. Username too long/short.
        '401':
          description: Unauthorized.
  /user/delete/:
    delete:
      tags:
        - Users
      summary: Delete a user.
      operationId: user.delete_user
      responses:
        '204':
          description: Successful operation.
        '401':
          description: Unauthorized.
...
components:
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
          example: Sky
        email:
          type: string
          description: This is only available when using the user's username/password in the authentication.
          example: sky@skystuff.cc
        super_admin:
          type: boolean
          description: If this field is set, projects should treat the user as an admin.
          example: true
        token:
          type: string
          description: The user's token, similar use to the hashed username/password. This is only available when using the user's username/password in the authentication.
...
    Project:
      type: object
      properties:
        name:
          type: string
          example: paste
        public:
          type: boolean
          description: This boolean is whether or not anybody can see this project. If it is false, users will be specifically added to it by an admin.
          example: false
        open:
          type: boolean
          description: This boolean is whether or not anybody can join this project. If it is false, users must be approved for the project.
          example: true
        description:
          type: string
          description: This is a generic description of the project. Markdown formatted.
    Projects:
      type: array
      items:
        $ref: '#/components/schemas/Project'
    ProjectStatus:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            approval:
              type: string
              description: The user's approval status for a project. One of; "approved", "denied", "pending", "inactive".
              example: inactive
    ProjectStatuses:
      type: array
      items:
        $ref: '#/components/schemas/ProjectStatus'
    Key:
      type: object
      properties:
        name:
          type: string
          description: Key name, maximum 256 characters.
          example: A generic API Key
        data:
          type: string
          description: Base64 (optionally) encoded data to store with the key.
          example: c3VwZXIgc2VjcmV0
        id:
          type: string
          description: The 128 random character ID of the key. This is the authorization for it.
        user:
          $ref: '#/components/schemas/User'
        project:
          $ref: '#/components/schemas/Project'
    Keys:
      type: array
      items:
        $ref: '#/components/schemas/Key'
...
    user.create:
      type: object
      properties:
        name:
          type: string
          description: Username of the user.
          example: Sky
        password:
          type: string
          description: Password of the user.
          example: AVerySecurePassword
        email:
          type: string
          description: Email of the user.
          example: sky@skystuff.cc
    user.login:
      type: object
      properties:
        name:
          type: string
          description: Username of the user.
          example: Sky
        password:
          type: string
          description: Password of the user.
          example: AVerySecurePassword
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: Authorization